style: move config to upper module
