+++ src/config/reader.rs
@@ -7,7 +7,7 @@
 use std::fs::File;
 use std::io::Read;
 
-use super::config::*;
+use super::Config;
 use crate::APP_ARGS;
 
 pub struct ConfigReader;
